+++ libcaf_core/test/save_inspector.cpp
@@ -809,4 +809,19 @@ end object)_";
   }
 }
 
+TEST_CASE("GH-1427 regression") {
+  struct opt_test {
+    std::optional<int> val;
+  };
+  auto x = opt_test{};
+  CHECK(f.object(x).fields(f.field("val", x.val).fallback(std::nullopt)));
+  CHECK(!f.get_error());
+  std::string baseline = R"_(
+begin object anonymous
+  begin optional field val
+  end field
+end object)_";
+  CHECK_EQ(f.log, baseline);
+}
+
 END_FIXTURE_SCOPE()