+++ CHANGELOG.md
@@ -45,6 +45,7 @@ is based on [Keep a Changelog](https://keepachangelog.com).
 - Fix a regression in `--dump-config` that caused CAF applications to emit
   malformed output.
 - Fix handling of WebSocket frames that are exactly on the 65535 byte limit.
+- Fix crash when using a fallback value for optional values (#1427).
 
 ## [0.19.2] - 2023-06-13